[Refactor] [4/N] Move VLLM_SERVER_DEV endpoints into the serve directory
